♻️ Move admin setup out of space (#1795)
