+++ apps/web/src/app/[locale]/admin-setup/make-me-admin-button.tsx
